# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural drawing management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [drawings]
      summary: List drawing categories
      responses:
        "200":
          description: Drawing categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [drawings]
      summary: Add a drawing category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [drawings]
      summary: Rename a drawing category
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [drawings]
      summary: Delete a drawing category
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all portal users
      responses:
        "200":
          description: Portal users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortalUser"
    post:
      operationId: adminCreateUser
      tags: [admin]
      summary: Create a portal user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created portal user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalUser"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a portal user
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated portal user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalUser"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a portal user
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "403":
          description: Administrator access required
  /admin/disciplines:
    get:
      operationId: adminListDisciplines
      tags: [admin]
      summary: List managed disciplines
      responses:
        "200":
          description: Managed disciplines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discipline"
    post:
      operationId: adminCreateDiscipline
      tags: [admin]
      summary: Add a drawing discipline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplineInput"
      responses:
        "201":
          description: Created discipline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discipline"
  /admin/disciplines/{id}:
    patch:
      operationId: adminUpdateDiscipline
      tags: [admin]
      summary: Rename a drawing discipline
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplineInput"
      responses:
        "200":
          description: Updated discipline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discipline"
    delete:
      operationId: adminDeleteDiscipline
      tags: [admin]
      summary: Delete a drawing discipline
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
...
  schemas:
...
    PortalUser:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        username:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, username, role, active, createdAt]
...
    AdminUserInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 4
        role:
          type: string
          enum: [admin, user]
      required: [name, username, password]
    AdminUserUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
    Discipline:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    DisciplineInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
    Category:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
...